Add nod_timing mode and fix resampler float precision bug
Adds VapGPT_nod_timing (single merged nod probability, HF auto-download
via maai-kyoto/vap_nod_timing_jp), fixes a float32 precision bug in
_CausalStreamingResampler, and fixes the checkpoint package loader so
nod/nod_timing state dicts actually get loaded instead of silently
staying at random init.

+++ src/maai/encoder.py
@@ -220,7 +220,7 @@ class _CausalStreamingResampler(nn.Module):
                 return x.new_zeros((batch_size, channels, 0))
 
             t = (
-                torch.arange(n_out, device=x.device, dtype=torch.float32) * float(step)
+                torch.arange(n_out, device=x.device, dtype=torch.float64) * float(step)
                 + float(self.state.next_t_in_samples)
             )
             t_floor = torch.floor(t)
